requirements.txt
streamlit
Pillow
ultralytics
torch
torchvision